feat: add null-ls
Refactors parts of the init.lua file and moves telescope into the new packer.lua file. The configs are now stored in the after plugin directory.
Some further migration is required as packer is getting into conflict when sourcing the config for the first time.
Null-LS enables some diagnostic and formatting. Autocompletion and go to definitions are still missing.